pyproject.toml
[tool.poetry]
name = "settleup_python_api"
version = "0.1.0"
description = ""
authors = ["kykpi <kylee.burgess.14@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
google-auth = "^1.28.0"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
